Rename tests.lib.test_case to tests.lib.anta
