--- tests/lib/parametrize.py
@@ -1,6 +1,0 @@
-# Copyright (c) 2023 Arista Networks, Inc.
-# Use of this source code is governed by the Apache License 2.0
-# that can be found in the LICENSE file.
-"""
-parametrize.py - Retrieves the mock data from the json_data file
-"""
