+++ tests/units/anta_tests/routing/test_ospf.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 from typing import Any
 
 from anta.tests.routing.ospf import VerifyOSPFNeighborCount, VerifyOSPFNeighborState
-from tests.lib.test_case import test  # noqa: F401
+from tests.lib.anta import test  # noqa: F401
 
 DATA: list[dict[str, Any]] = [
     {
