+++ tests/units/anta_tests/test_field_notices.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 from typing import Any
 
 from anta.tests.field_notices import VerifyFieldNotice44Resolution, VerifyFieldNotice72Resolution
-from tests.lib.test_case import test  # noqa: F401
+from tests.lib.anta import test  # noqa: F401
 
 DATA: list[dict[str, Any]] = [
     {
