+++ tests/units/anta_tests/test_hardware.py
@@ -15,7 +15,7 @@ from anta.tests.hardware import (
     VerifyTransceiversManufacturers,
     VerifyTransceiversTemperature,
 )
-from tests.lib.test_case import test  # noqa: F401
+from tests.lib.anta import test  # noqa: F401
 
 DATA: list[dict[str, Any]] = [
     {
