+++ tests/units/anta_tests/test_mlag.py
@@ -1,3 +1,6 @@
+# Copyright (c) 2023 Arista Networks, Inc.
+# Use of this source code is governed by the Apache License 2.0
+# that can be found in the LICENSE file.
 """
 Tests for anta.tests.mlag.py
 """
@@ -6,7 +9,7 @@ from __future__ import annotations
 from typing import Any
 
 from anta.tests.mlag import VerifyMlagConfigSanity, VerifyMlagDualPrimary, VerifyMlagInterfaces, VerifyMlagReloadDelay, VerifyMlagStatus
-from tests.lib.test_case import test  # noqa: F401
+from tests.lib.anta import test  # noqa: F401
 
 DATA: list[dict[str, Any]] = [
     {
